requirements.txt
streamlit==1.33.0
streamlit-drawable-canvas==0.9.2
tensorflow==2.15.0
numpy==1.24.3
pillow==9.5.0
opencv-python==4.8.0.74
